+++ tests/migrations/helpers.py
@@ -4,10 +4,7 @@ This module contains various helpers for migrations testing.
 """
 import os
 
-from django import VERSION, setup
-
-
-setup()
+from django import VERSION
 
 
 MIGRATION_NAME = 'test'
@@ -16,6 +13,10 @@ MIGRATION_NAME = 'test'
 def makemigrations():
     from django.core.management import call_command
     from django.core.management.commands.makemigrations import Command
+    from django.db.migrations import questioner
+
+    # We should answer yes for all migrations questioner questions
+    questioner.input = lambda x: 'y'
 
     os.system('find . -name \*.pyc -delete')
     if VERSION >= (1, 10):